# openapi.yaml
openapi: 3.1.0
info:
  title: EOS Visitor Management API
  version: 0.1.0
  description: |
    Backend for the EOS visitor management system.

    Two security zones:
      * `/api/v1/public/**` — reached by visitors from the web app opened via the
        location QR code. No OIDC. A visit is authorized by the opaque
        `visitToken` handed out at registration.
      * everything else — admin panel, OIDC bearer token, role `eos-admin`.

    EOS records **visits, not visitors.** There is no visitor entity and no
    recognition across visits: contact details belong to the single visit they
    were given for, and are purged with it per the location's `retentionDays`.
    `selfCheckIn` therefore never recognises a returning visitor: this endpoint is
    unauthenticated, and `visitorName` plus `contact.email` are not secrets, so
    reissuing an existing `visitToken` on a match would hand a guessable pair of
    facts the power to read someone's badge and check them out. A duplicate
    registration is instead resolved at the reception desk.

    `badgeUrl` and the 201 `Location` header carry the `visitId` only — never the
    `visitToken`. The token is returned exactly once, in the response body, and is
    replayed by the visitor app as `X-Visit-Token`. Keeping it out of URLs keeps it
    out of access logs, browser history and `Referer` headers.
servers:
  - url: /
...
paths:
...
  /api/v1/public/visits/{visitId}:
    get:
      tags: [Visitor Self-Service]
      summary: Digital badge state
      operationId: getOwnVisit
      security: []
      parameters:
        - $ref: '#/components/parameters/VisitId'
        - $ref: '#/components/parameters/VisitTokenHeader'
      responses:
        '200':
          description: Badge data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VisitBadge' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Visit purged by retention policy
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/v1/visits/{visitId}:
    get:
      tags: [Visits]
      summary: Get one visit
      operationId: getVisit
      parameters:
        - $ref: '#/components/parameters/VisitId'
      responses:
        '200':
          description: Visit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Visits]
      summary: Correct visit data
      operationId: updateVisit
      parameters:
        - $ref: '#/components/parameters/VisitId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/VisitUpdateRequest' }
      responses:
        '200':
          description: Updated visit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    delete:
      tags: [Visits]
      summary: Delete a visit (GDPR erasure)
      operationId: deleteVisit
      parameters:
        - $ref: '#/components/parameters/VisitId'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://example.invalid/realms/eos/.well-known/openid-configuration
  parameters:
    VisitId:
      name: visitId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    VisitTokenHeader:
      name: X-Visit-Token
      in: header
      required: true
      description: Opaque token returned by `selfCheckIn`.
      schema: { type: string, minLength: 32, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag of the version being modified, for optimistic locking.
      schema: { type: string }
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but lacking the required role
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: ETag mismatch — the resource changed in the meantime
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionRequired:
      description: The `If-Match` header is required but was not supplied
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    VisitStatus:
      type: string
      enum: [EXPECTED, ON_SITE, CHECKED_OUT]
      description: |
        * `EXPECTED` — pre-registered by an admin, not yet on site
        * `ON_SITE` — checked in, not checked out (`abgemeldet = false`)
        * `CHECKED_OUT` — checked out (`abgemeldet = true`)
    ContactInfo:
      type: object
      description: |
        `text` is the contact detail exactly as the visitor typed it. It is never
        parsed — only displayed and exported — so it accommodates real-world input
        such as `030 12345 extension 7`.

        `email` is populated only when the visitor uses the mail field. It is the
        only machine-usable contact field and the only one `selfCheckIn` uses to
        recognise a returning visitor.
      required: [text]
      properties:
        text: { type: string, minLength: 3, maxLength: 200 }
        email: { type: string, format: email, maxLength: 254 }
    Address:
      type: object
      required: [street, postalCode, city]
      properties:
        street: { type: string, maxLength: 120 }
        postalCode: { type: string, maxLength: 12 }
        city: { type: string, maxLength: 80 }
        country:
          type: string
          description: ISO 3166-1 alpha-2.
          pattern: '^[A-Z]{2}$'
          default: DE
    PublicLocation:
      type: object
      required: [id, companyName, address]
      properties:
        id: { type: string, format: uuid }
        companyName: { type: string }
        address: { $ref: '#/components/schemas/Address' }
        additionalInfo:
          type: string
          description: Free text shown on the registration form (Zusatz Info).
        hostRequired:
          type: boolean
          default: false
        privacyNoticeUrl: { type: string, format: uri }
...
    Visit:
      type: object
      required: [id, locationId, visitorName, visitDate, purpose, status, contact, createdAt]
      properties:
        id: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        visitorName: { type: string, maxLength: 120 }
        visitorCompany: { type: string, maxLength: 120 }
        visitDate:
          type: string
          format: date
          description: Day the visit is registered for (Datum).
        purpose:
          type: string
          maxLength: 500
          description: Reason for the visit (Anlass).
        hostName:
          type: string
          maxLength: 120
          description: Employee being visited.
        contact: { $ref: '#/components/schemas/ContactInfo' }
        status: { $ref: '#/components/schemas/VisitStatus' }
        checkInAt: { type: ['string', 'null'], format: date-time }
        checkOutAt: { type: ['string', 'null'], format: date-time }
        selfRegistered: { type: boolean }
        createdAt: { type: string, format: date-time }
        createdBy:
          type: ['string', 'null']
          description: Admin subject that created the visit; null for self-registration.
        updatedAt: { type: string, format: date-time }
    VisitBadge:
      type: object
      description: Visitor-facing projection — no admin or cross-visit data.
      required: [id, visitorName, visitDate, status, location]
      properties:
        id: { type: string, format: uuid }
        visitorName: { type: string }
        visitorCompany: { type: string }
        visitDate: { type: string, format: date }
        purpose: { type: string }
        hostName: { type: string }
        status: { $ref: '#/components/schemas/VisitStatus' }
        checkInAt: { type: ['string', 'null'], format: date-time }
        checkOutAt: { type: ['string', 'null'], format: date-time }
        location: { $ref: '#/components/schemas/PublicLocation' }
...
    VisitUpdateRequest:
      type: object
      description: Merge patch. Only the listed fields are mutable; status changes go through the check-in/check-out endpoints.
      properties:
        visitorName: { type: string, minLength: 1, maxLength: 120 }
        visitorCompany: { type: ['string', 'null'], maxLength: 120 }
        visitDate: { type: string, format: date }
        purpose: { type: string, minLength: 1, maxLength: 500 }
        hostName: { type: ['string', 'null'], maxLength: 120 }
        contact: { $ref: '#/components/schemas/ContactInfo' }
...
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          description: Field-level validation failures.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
security:
  - oidc: []